+++ app/src/main/java/com/hugopinto/srn/Fragments/Agregar.java
@@ -90,7 +90,7 @@ public class Agregar extends Fragment {
         btn.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View v){
-                if(Integer.parseInt(not.getText().toString())>10){
+                if(Float.parseFloat(not.getText().toString())>10){
                     Toast.makeText(v.getContext(), "Nota mayor a 10, ingrese una nota menor o igual a 10", Toast.LENGTH_SHORT).show();
                 }else {
                     boolean flag = DBHelper.myDB.add(new Persona(carn.getText().toString(), not.getText().toString(), mat.getText().toString(), cat.getText().toString()));
